@@ -1,9 +1,9 @@
 {
     "name": "kevinpapst/lockdown-per-user-plugin",
-    "description": "A Kimai plugin that allows to set the lockdown period per user.",
-    "homepage": "https://www.kimai.org/store/translation-bundle.html",
+    "description": "Allows to set the lockdown period per user.",
+    "homepage": "https://www.kevinpapst.de",
     "type": "kimai-plugin",
-    "version": "1.0",
+    "version": "2.0",
     "keywords": [
         "kimai",
         "kimai-plugin"
@@ -18,8 +18,8 @@
     ],
     "extra": {
         "kimai": {
-            "require": "1.16.9",
-            "name": "LockdownPerUserBundle"
+            "require": 20000,
+            "name": "Lockdown per user"
         }
     },
     "autoload": {
@@ -33,7 +33,7 @@
             "symfony/flex": false
         },
         "platform": {
-            "php": "7.3"
+            "php": "8.1.3"
         },
         "preferred-install": {
             "*": "dist"
@@ -51,18 +51,14 @@
             "@phpstan"
         ]
     },
-    "require": {
-        "doctrine/dbal": "^2.1"
-    },
+    "minimum-stability": "dev",
     "require-dev": {
-        "doctrine/doctrine-fixtures-bundle": "^3.3",
-        "fakerphp/faker": "^1.15",
         "friendsofphp/php-cs-fixer": "^3.0",
-        "kevinpapst/kimai2": "^1.15",
+        "kevinpapst/kimai2": "^2.0",
         "phpstan/phpstan": "^1.0",
         "phpstan/phpstan-doctrine": "^1.0",
         "phpstan/phpstan-symfony": "^1.0",
-        "symfony/console": "^4.0",
-        "symfony/event-dispatcher": "^4.0"
+        "symfony/console": "^6.0",
+        "symfony/event-dispatcher": "^6.0"
     }
 }
